Dodana implementacija servisa

+++ eTeatar/eTeatar.Model/Hrana.cs
@@ -10,7 +10,7 @@ namespace eTeatar.Model
 
         public string Naziv { get; set; }
 
-        public float Cijena { get; set; }
+        public decimal Cijena { get; set; }
 
         public virtual ICollection<StavkaUplate> StavkaUplates { get; set; } = new List<StavkaUplate>();
     }
